Move public models into Models folder
