+++ Sources/WKData/Models/WKPageWatchStatus.swift
@@ -1,0 +1,4 @@
+public struct WKPageWatchStatus {
+    public let watched: Bool
+    public let userHasRollbackRights: Bool?
+}
