+++ Sources/WKData/Models/WKUndoOrRollbackResult.swift
@@ -1,0 +1,4 @@
+public struct WKUndoOrRollbackResult: Codable {
+    public let newRevisionID: Int
+    public let oldRevisionID: Int
+}
