+++ Sources/WKData/Models/WKWatchlist.swift
@@ -1,0 +1,21 @@
+import Foundation
+
+public struct WKWatchlist {
+    
+    public struct Item {
+        public let title: String
+        public let revisionID: UInt
+        public let oldRevisionID: UInt
+        public let username: String
+        public let isAnon: Bool
+        public let isBot: Bool
+        public let timestamp: Date
+        public let commentWikitext: String
+        public let commentHtml: String
+        public let byteLength: UInt
+        public let oldByteLength: UInt
+        public let project: WKProject
+    }
+    
+    public let items: [Item]
+}
