+++ Sources/WKData/Models/WKWatchlistExpiryType.swift
@@ -1,0 +1,7 @@
+public enum WKWatchlistExpiryType: String {
+    case never
+    case oneWeek = "1 week"
+    case oneMonth = "1 month"
+    case threeMonths = "3 months"
+    case sixMonths = "6 months"
+}
